Exclude .then() and .catch() from automatic handling.
By default, all symbols except .hasOwnProperty() are stubbed even if
they are not assigned a value. This is to be able to track references to
all invocations. However, the way that Promise.resolve() works, it
checks for the presence of a .then() property and if it is there assumes
that the resolve value is a promise so it tries to get its resolved
value. But this is a mock reference so the stub does not return a
resolved value so the promise never resolves.

This fix excludes .then() and .catch() from automatically creating a
stub so that Promise will not try to resolve it, but the user can still
override the call with a when() clause.

This fix also adds "Symbol(Symbol.toPrimitive)" to the exclude list so
that by default toString() on a mock will return something other than
null.